made postauth view and tabs for postauth us to add to eihter customer view or seller view
