+++ src/main/java/com/oop/BargainBin/AccountType.java
@@ -10,4 +10,6 @@ package com.oop.BargainBin;
  */
 public interface AccountType {
     
+    public void loadUserPlatform();
+    
 }
