+++ src/main/java/com/oop/BargainBin/BargainBin.java
@@ -105,4 +105,5 @@ public class BargainBin {
     private LoginRegisterView logRegView;
     private AccountModel accModel;
     
+    
 }
