+++ src/main/java/com/oop/BargainBin/CustomerAccount.java
@@ -8,9 +8,14 @@ package com.oop.BargainBin;
  *
  * @author binup
  */
-public class CustomerAccount {
+public class CustomerAccount implements AccountType{
 
     CustomerAccount(AccountModel aThis) {
+        loadUserPlatform();
+    }
+
+    @Override
+    public void loadUserPlatform() {
         throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
     }
