+++ src/main/java/com/oop/BargainBin/SellerAccount.java
@@ -8,9 +8,14 @@ package com.oop.BargainBin;
  *
  * @author binup
  */
-public class SellerAccount extends AccountController {
+public class SellerAccount implements AccountType{
 
     SellerAccount(AccountModel aThis) {
+        loadUserPlatform();
+    }
+
+    @Override
+    public void loadUserPlatform() {
         throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
     }
